quotes/requirements.txt
django>=4.2
psycopg2-binary>=2.9
pillow>=10.0
mongoengine>=0.24.0
python-dateutil>=2.8.2
pymongo>=4.0.0
python-decouple>=3.8